source/frontend/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap4==4.0.2
Flask-WTF==1.0.0
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==1.44.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
protobuf==3.19.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.2
requests==2.27.0
rsa==4.8
six==1.16.0
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.1
